Add getById method to StudentController and handling cases when entity by id was not found in related methods.

+++ contractors/src/main/java/com/university/contractors/controller/StudentController.java
@@ -1,11 +1,16 @@
 package com.university.contractors.controller;
 
+import com.university.contractors.config.Endpoints;
 import com.university.contractors.model.Student;
 import com.university.contractors.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
-@RestController(value = "/students")
+import java.util.Optional;
+
+@RestController
 public class StudentController {
 
     private final StudentRepository studentRepository; // todo add service layer
@@ -15,23 +20,40 @@ public class StudentController {
         this.studentRepository = studentRepository;
     }
 
-    @GetMapping
+    @GetMapping(path = Endpoints.STUDENT_BY_ID)
+    public Student getById(@PathVariable Long id) {
+        Optional<Student> optionalStudentById = studentRepository.findById(id);
+
+        if (optionalStudentById.isPresent()) {
+            return optionalStudentById.get();
+        }
+
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Student with id '%s' was not found.");
+    }
+
+    @GetMapping(path = Endpoints.STUDENTS)
     public Iterable<Student> getAll() {
         return studentRepository.findAll();
     }
 
-    @PostMapping
+    @PostMapping(path = Endpoints.STUDENTS)
     public Student create(@RequestBody Student student) {
         return studentRepository.save(student);
     }
 
-    @PutMapping
+    @PutMapping(path = Endpoints.STUDENTS)
     public Student update(@RequestBody Student student) {
         return studentRepository.save(student);
     }
 
-    @DeleteMapping
-    public void delete(@RequestParam Long id) {
-        studentRepository.deleteById(id);
+    @DeleteMapping(path = Endpoints.STUDENT_BY_ID)
+    public void delete(@PathVariable Long id) {
+        boolean isStudentWithGivenIdExists = studentRepository.existsById(id);
+
+        if (isStudentWithGivenIdExists) {
+            studentRepository.deleteById(id);
+        }
+
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Student with id '%s' was not found.");
     }
 }
